+++ src/main/java/io/aiven/elasticsearch/repositories/metadata/EncryptionKeyMetadata.java
@@ -1,0 +1,45 @@
+/*
+ * Copyright 2020 Aiven Oy
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.aiven.elasticsearch.repositories.metadata;
+
+import com.fasterxml.jackson.annotation.JsonAutoDetect;
+import com.fasterxml.jackson.annotation.JsonCreator;
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+@JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
+class EncryptionKeyMetadata {
+
+    private final String key;
+
+    private final int version;
+
+    @JsonCreator
+    public EncryptionKeyMetadata(@JsonProperty("key") final String key,
+                                 @JsonProperty("version") final int version) {
+        this.key = key;
+        this.version = version;
+    }
+
+    public String key() {
+        return key;
+    }
+
+    public int version() {
+        return version;
+    }
+
+}
