--- src/main/java/io/aiven/elasticsearch/storage/security/Encryption.java
@@ -1,49 +1,0 @@
-/*
- * Copyright 2020 Aiven Oy
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.aiven.elasticsearch.storage.security;
-
-import javax.crypto.Cipher;
-import javax.crypto.NoSuchPaddingException;
-
-import java.security.InvalidKeyException;
-import java.security.Key;
-import java.security.NoSuchAlgorithmException;
-import java.security.SecureRandom;
-import java.util.Objects;
-
-import org.bouncycastle.jce.provider.BouncyCastleProvider;
-
-public interface Encryption {
-
-    default Cipher createEncryptingCipher(final Key key,
-                                          final String transformation) {
-        Objects.requireNonNull(key, "key hasn't been set");
-        Objects.requireNonNull(transformation, "transformation hasn't been set");
-        try {
-            final var cipher = Cipher.getInstance(transformation, new BouncyCastleProvider());
-            cipher.init(
-                    Cipher.ENCRYPT_MODE,
-                    key,
-                    SecureRandom.getInstanceStrong());
-            return cipher;
-        } catch (final NoSuchAlgorithmException | NoSuchPaddingException
-                | InvalidKeyException e) {
-            throw new RuntimeException("Couldn't create encrypt cipher", e);
-        }
-    }
-
-}
