+++ src/main/resources/plugin-descriptor.properties
@@ -26,7 +26,7 @@ version=${version}
 name=aiven-repository-gcs
 #
 # 'classname': the name of the class to load, fully-qualified.
-classname=io.aiven.elasticsearch.gcs.GcsStoragePlugin
+classname=io.aiven.elasticsearch.repositories.gcs.GcsRepositoryPlugin
 #
 # 'java.version': version of java the code is built against
 # use the system property java.specification.version
